Derive AWS Runner from Out-of-Process Runner
