+++ rastervision/runner/__init__.py
@@ -3,6 +3,7 @@
 from rastervision.runner.command_definition import *
 from rastervision.runner.command_dag import *
 from rastervision.runner.experiment_runner import *
+from rastervision.runner.out_of_process_experiment_runner import *
 from rastervision.runner.inprocess_experiment_runner import *
 from rastervision.runner.aws_batch_experiment_runner import *
 from rastervision.runner.local_experiment_runner import *
