+++ rastervision/runner/out_of_process_experiment_runner.py
@@ -1,0 +1,80 @@
+import os
+import click
+
+from rastervision.runner import ExperimentRunner
+from rastervision.utils.files import save_json_config
+from rastervision.cli import Verbosity
+
+
+def make_command(command_config_uri, tmp_dir=None):
+    verbosity = Verbosity.get()
+    v_flag = 'v' * max(0, verbosity - 1)
+    if v_flag:
+        v_flag = '-{}'.format(v_flag)
+    if tmp_dir is None:
+        command = 'python -m rastervision {} run_command {}'.format(
+            v_flag, command_config_uri)
+    else:
+        command = 'python -m rastervision {} run_command {} --tempdir {}'.format(
+            v_flag, command_config_uri, tmp_dir)
+    return command
+
+
+class OutOfProcessExperimentRunner(ExperimentRunner):
+    def __init__(self):
+        self.tmp_dir = None
+
+    def _run_experiment(self, command_dag):
+        """Runs all commands."""
+
+        ids_to_job = {}
+        for command_id in command_dag.get_sorted_command_ids():
+            command_def = command_dag.get_command_definition(command_id)
+            command_config = command_def.command_config
+            command_root_uri = command_config.root_uri
+            command_uri = os.path.join(command_root_uri, 'command-config.json')
+            print('Saving command configuration to {}...'.format(command_uri))
+            save_json_config(command_config.to_proto(), command_uri)
+
+            parent_job_ids = []
+            for upstream_id in command_dag.get_upstream_command_ids(
+                    command_id):
+                if upstream_id not in ids_to_job:
+                    cur_command = (command_config.command_type, command_id)
+                    u = command_dag.get_command(upstream_id)
+                    upstream_command = (u.command_type, upstream_id)
+                    raise Exception(
+                        '{} command has parent command of {}, '
+                        'but does not exist in previous batch submissions - '
+                        'topological sort on command_dag error.'.format(
+                            cur_command, upstream_command))
+                parent_job_ids.append(ids_to_job[upstream_id])
+
+            batch_run_command = make_command(command_uri, self.tmp_dir)
+            job_id = self.submit(
+                command_config.command_type,
+                command_def.experiment_id,
+                self.job_queue,
+                self.job_definition,
+                batch_run_command,
+                attempts=self.attempts,
+                parent_job_ids=parent_job_ids)
+
+            ids_to_job[command_id] = job_id
+
+    def _dry_run(self, command_dag):
+        """Runs all commands."""
+        click.echo(
+            click.style(
+                '\n{} commands to be issued:'.format(
+                    self.execution_environment),
+                fg='green',
+                bold=True,
+                underline=True))
+        for command_id in command_dag.get_sorted_command_ids():
+            command_def = command_dag.get_command_definition(command_id)
+            command_config = command_def.command_config
+            command_root_uri = command_config.root_uri
+            command_uri = os.path.join(command_root_uri, 'command-config.json')
+            batch_run_command = make_command(command_uri, self.tmp_dir)
+            click.echo('  {}'.format(batch_run_command))
